init req.flash

+++ api/controllers/articleController.js
@@ -14,28 +14,52 @@ module.exports = {
     },
     post: async(req, res) => {
         const dbArticle = await Article.find({})
-        Article.create({
+        if (req.body.title) {
+            Article.create({
                 ...req.body
-            },
-            res.redirect('/article')
-        )
+            })
+            req.flash('success', 'Merci d avoir entrer un titre')
+            res.render('article', {
+                success: req.flash('success'),
+                dbArticle: dbArticle
+            })
+
+        } else {
+            req.flash('errors', 'Vous devez saisir un titre')
+            res.render('article', {
+                errors: req.flash('errors'),
+                dbArticle: dbArticle
+            })
+        }
     },
-    deleteOne: (req, res) => {
-        Article.deleteOne({
+    deleteOne: async(req, res) => {
+        const dbArticle = await Article.find({})
+        if (req.params.id) {
+            Article.deleteOne({
                 _id: req.params.id
             },
             (err) => {
                 if (!err) {
-                    res.redirect('/article')
+                    req.flash('success', 'Suppression confirmer !')
+                    res.render('article', {
+                        success: req.flash('success'),
+                        dbArticle: dbArticle
+                    })
                 } else {
                     res.send(err)
                 }
             })
+        } else req.flash('errors', 'Une erreur c est produit !'), res.render('article', {errors: req.flash('errors')})
     },
-    deleteAll: (req, res) => {
+    deleteAll: async(req, res) => {
+        const dbArticle = await Article.find({})
         Article.deleteMany((err) => {
             if (!err) {
-                res.redirect('/article')
+                req.flash('success', 'Tous les articles ont été supprimer !')
+                res.render('article', {
+                    success: req.flash('success'),
+                    dbArticle: dbArticle
+                })
             } else {
                 res.send(err)
             }
